worker/requirements.txt
Flask==2.2.2
Werkzeug<3.0.0
redis==4.5.1
psycopg2-binary==2.9.5